+++ src/points/mod.rs
@@ -6,7 +6,7 @@ pub use self::parser::*;
 /// Representation of the [`<list-of-points>`] type.
 ///
 /// [`<list-of-points>`]: https://www.w3.org/TR/SVG11/shapes.html#PointsBNF
-#[derive(Clone, PartialEq)]
+#[derive(Clone, PartialEq, Default)]
 pub struct Points(pub Vec<(f64, f64)>);
 
 impl_from_vec!(Points, Points, (f64, f64));